borrowing books

+++ src/main/java/com/system/libsystem/entities/user/UserEntity.java
@@ -1,5 +1,6 @@
 package com.system.libsystem.entities.user;
 
+import com.system.libsystem.util.UserType;
 import lombok.Getter;
 import lombok.Setter;
 import org.springframework.security.core.GrantedAuthority;
@@ -16,7 +17,7 @@ import java.util.List;
 public class UserEntity implements UserDetails {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
+    private int id;
 
     @Column(nullable = false, unique = true)
     private Long cardNumber;
